implement overlapping tile removal

+++ src/main.cc
@@ -28,7 +28,7 @@ int main(int argc, char* argv[]) {
 	}
 
 	for(size_t i = 0; i < container.getChunkCount(); i++) {
-		for(size_t egg = 0; egg < 15; egg++) {
+		for(int egg = 0; egg < 15; egg++) {
 			Chunk &chunk = container.getChunk(i);
 			OverlappingTile* alien = new OverlappingTile();
 			alien->setPosition(glm::uvec3(egg + chunk.getPosition().x * Chunk::Size, egg + chunk.getPosition().y * Chunk::Size, chunk.height - 1));
